Upgrade Fixes
  - fix bug in bash native credential parsing functions
  - fix resetting of credentials in v0.73 upgrade scripts
  - fix bug where multiple scripts could clobber `apt.conf.d` settings
  - add dpkg lock timeout to allow other processes to release apt lock
  - fix kamailio config not regenerated on upgrade
  - add more robust resetting of config files on upgrade failure
  - fix centos7 /etc/default/kamailio.conf in wrong place
  - fix upgrade log is not cleared when clicking show previous log
  - fix typo in `RESTART_DAEMONIZE` variable
  - fix cursor not a pointer on new reload buttons
  - fix reload lib functions to use new reload buttons
  - fix upgrade from gui hanging

+++ kamailio/debian/11.sh
@@ -22,6 +22,11 @@ function install() {
     apt-get install -y curl wget sed gawk vim perl uuid-dev libssl-dev logrotate rsyslog \
         libcurl4-openssl-dev libjansson-dev cmake firewalld python3 build-essential python3-venv
 
+    if (( $? != 0 )); then
+        printerr 'Failed installing required packages'
+        return 1
+    fi
+
     # we need a newer version of certbot than the distro repos offer
     apt-get remove -y *certbot*
     python3 -m venv /opt/certbot/
@@ -122,8 +127,8 @@ EOF
     firewall-cmd --reload
 
     # Configure Kamailio systemd service
-    cp -f ${DSIP_PROJECT_DIR}/kamailio/systemd/kamailio-v2.service /etc/systemd/system/kamailio.service
-    chmod 644 /etc/systemd/system/kamailio.service
+    cp -f ${DSIP_PROJECT_DIR}/kamailio/systemd/kamailio-v2.service /lib/systemd/system/kamailio.service
+    chmod 644 /lib/systemd/system/kamailio.service
     systemctl daemon-reload
     systemctl enable kamailio
